+++ src/controllers/commands/builtin/twitch/SendWhisper.cpp
@@ -8,9 +8,7 @@
 #include "messages/Message.hpp"
 #include "messages/MessageBuilder.hpp"
 #include "messages/MessageElement.hpp"
-#include "providers/bttv/BttvEmotes.hpp"
 #include "providers/emoji/Emojis.hpp"
-#include "providers/ffz/FfzEmotes.hpp"
 #include "providers/twitch/api/Helix.hpp"
 #include "providers/twitch/TwitchAccount.hpp"
 #include "providers/twitch/TwitchIrcServer.hpp"
@@ -107,8 +105,7 @@ bool appendWhisperMessageWordsLocally(const QStringList &words)
 
     const auto &acc = app->getAccounts()->twitch.getCurrent();
     const auto &accemotes = *acc->accessEmotes();
-    const auto *bttvemotes = app->getBttvEmotes();
-    const auto *ffzemotes = app->getFfzEmotes();
+    const auto *emoteController = app->getEmotes();
     auto flags = MessageElementFlags();
     auto emote = std::optional<EmotePtr>{};
     for (int i = 2; i < words.length(); i++)
@@ -122,21 +119,16 @@ bool appendWhisperMessageWordsLocally(const QStringList &words)
             }
         }  // Twitch emote
 
-        {  // bttv/ffz emote
-            emote = bttvemotes->emote({words[i]});
-            if (!emote)
-            {
-                emote = ffzemotes->emote({words[i]});
-            }
-            // TODO: Load 7tv global emotes
+        {  // third party emotes
+            emote = emoteController->resolveGlobal(EmoteName{words[i]});
             if (emote)
             {
                 b.emplace<EmoteElement>(*emote, MessageElementFlag::Emote);
                 continue;
             }
-        }  // bttv/ffz emote
+        }  // third party emotes
         {  // emoji/text
-            for (auto &variant : app->getEmotes()->getEmojis()->parse(words[i]))
+            for (auto &variant : emoteController->getEmojis()->parse(words[i]))
             {
                 constexpr const static struct {
                     void operator()(EmotePtr emote, MessageBuilder &b) const