+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Dev: Correct handling of eventsubs without any account. (#6503)
 - Dev: Removed dependency to Qt5 Compatibility module by updating libcommuni. (#6500)
 - Dev: Merged emote element flags from different providers into two. (#6511)
+- Dev: Refactored thrid party emotes. (#6436)
 - Dev: Fixed warnings on Qt 6.10. (#6422)
 - Dev: Removed unused method in `Emojis`. (#6517)
 - Dev: Refactored `Emotes` into `EmoteController`. (#6516)